Move Claims type into custom_service_account
