+++ src/lib.rs
@@ -95,7 +95,6 @@ mod default_authorized_user;
 mod default_service_account;
 mod error;
 mod gcloud_authorized_user;
-mod jwt;
 mod types;
 mod util;
